MOVE-4884 ta ut kotlinX modelen i eget module
