+++ nhn-model/pom.xml
@@ -1,0 +1,133 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>no.difi.meldingsutveksling.nhn-adapter</groupId>
+        <artifactId>parent-pom</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>nhn-model</artifactId>
+    <packaging>jar</packaging>
+    <name>NHN transport model</name>
+
+    
+    <scm>
+        <connection>scm:git:https://github.com/felleslosninger/efm-nhn-adapter</connection>
+        <developerConnection>scm:git:https://github.com/felleslosninger/efm-nhn-adapter</developerConnection>
+        <url>https://github.com/felleslosninger/efm-nhn-adapter</url>
+    </scm>
+
+    <licenses>
+        <license>
+            <name>The Apache License, Version 2.0</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+        </license>
+    </licenses>
+
+    <developers>
+        <developer>
+            <name>Digdir</name>
+            <email>servicedesk@digdir.no</email>
+            <organization>Digdir</organization>
+            <organizationUrl>https://www.digdir.no</organizationUrl>
+        </developer>
+    </developers>
+
+    <properties>
+        <java.version>21</java.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <helseid.version>1.0.1</helseid.version>
+        <msh-client.version>1.4.0</msh-client.version>
+        <kotlin.version>2.2.0</kotlin.version>
+        <kotest.version>5.9.1</kotest.version>
+        <kotest-spring.version>1.3.0</kotest-spring.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-serialization-json-jvm</artifactId>
+            <version>1.9.0</version>
+        </dependency>
+        <dependency>
+            <groupId>no.ks.fiks</groupId>
+            <artifactId>msh-client</artifactId>
+            <version>${msh-client.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-serialization-core-jvm</artifactId>
+            <version>1.9.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-datetime-jvm</artifactId>
+            <version>0.7.1</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlin</groupId>
+            <artifactId>kotlin-stdlib-jdk8</artifactId>
+            <version>${kotlin.version}</version>
+        </dependency>
+       
+    </dependencies>
+    <build>
+        <sourceDirectory>src/main/kotlin</sourceDirectory>
+          <plugins>
+              <plugin>
+              <groupId>org.jetbrains.kotlin</groupId>
+              <artifactId>kotlin-maven-plugin</artifactId>
+                  <executions>
+                      <execution>
+                          <id>compile</id>
+                          <phase>compile</phase>
+                          <goals><goal>compile</goal></goals>
+                          <configuration>
+                              <sourceDirs>
+                                  <sourceDir>${project.basedir}/src/main/kotlin</sourceDir>
+                              </sourceDirs>
+                          </configuration>
+                      </execution>
+                      <execution>
+                          <id>test-compile</id>
+                          <phase>test-compile</phase>
+                          <goals>
+                              <goal>test-compile</goal>
+                          </goals>
+                          <configuration>
+                              <sourceDirs>
+                                  <sourceDir>${project.basedir}/src/test/kotlin</sourceDir>
+                              </sourceDirs>
+                          </configuration>
+                      </execution>
+                  </executions>
+
+              <configuration>
+                  <jvmTarget>21</jvmTarget>
+                  <compilerPlugins>
+                      <plugin>kotlinx-serialization</plugin>
+                  </compilerPlugins>
+              </configuration>
+
+              <dependencies>
+                  <dependency>
+                      <groupId>org.jetbrains.kotlin</groupId>
+                      <artifactId>kotlin-maven-serialization</artifactId>
+                      <version>${kotlin.version}</version>
+                  </dependency>
+                  <dependency>
+                      <groupId>no.ks.fiks</groupId>
+                      <artifactId>msh-client</artifactId>
+                      <version>${msh-client.version}</version>
+                  </dependency>
+              </dependencies>
+
+
+              </plugin>
+        </plugins>
+    </build>
+
+</project>
