+++ nhn-model/src/main/kotlin/no/difi/meldingsutveksling/nhn/adapter/model/MessageStatus.kt
